Update ruby gemspec for Rails 4 support (#151)
* Update ruby gemspec for rails 4 support

Safcoin are the first MTO to integrate using our Ruby SDK, but they're
using Rails 4 which causes conflicts with the runtime dependency
versions listed in the generated gemspec. They have been able to clone
and update the gem for now, but this will make it harder for them to
upgrade it later if/when we add new features.

This change updates the gemspec.mustache file to require lower versions
of the json and activesupport runtime dependencies.

The change has been tested with both Rails 4 and Rails 5 apps, with both
successfully able to make API requests using this lower requirement.

* checkin open-api.json changes for review