Upgrade the PHP api tempalte to the 4.0.0-beta3 version (#169)
It looks like we accidentally had the old swagger 3 version generator for the PHP api, which doesn't support x-group-params. Note that the documentation we genereated however was for the 4.0.0-beta3 version, and we had that in our main docs as well.

This change fixes this, and upgrades the template to the proper version. Note that this will generate a PHP SDK which is partially incompatible with the older versions so our customers have to be told!