BTA-3427 Ensure the request body is the same as the data for the request signature (#193)
TradeBFX requests fail randomly with an Authorzation error (401). The validity of the signature
depends on the http method, url, body, api_key and api_secret being correct. Of these the only
parameter which could vary between requests is the body and this commit ensures the same body param is used to sign the request and in the API call.
It also adds the request options to the error logging so we have all we need to debug API errors
in HoneyBadger going forward.

Co-authored-by: Ekerete <>